cleanup packages, memoize
