+++ package.json
@@ -22,14 +22,12 @@
     "@types/react": "^16.9.0",
     "@types/react-dom": "^16.9.0",
     "classnames": "^2.2.6",
-    "cloudinary-react": "^1.6.7",
     "js-search": "^2.0.0",
     "lodash": "^4.17.19",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.1",
-    "react-virtualized": "^9.22.2",
     "swr": "^0.3.8",
     "typescript": "~3.7.2"
   },
@@ -61,7 +59,6 @@
     "@types/js-search": "^1.4.0",
     "@types/lodash": "^4.14.157",
     "@types/react-router-dom": "^5.1.5",
-    "@types/react-virtualized": "^9.21.10",
     "autoprefixer": "^9.8.4",
     "chromatic": "^5.0.0",
     "postcss-cli": "^7.1.1",
